requirements-docker.txt
beautifulsoup4
colorama==0.4.6
openai==0.27.2
playsound==1.2.2
python-dotenv==1.0.0
pyyaml==6.0
readability-lxml==0.8.1
requests
tiktoken==0.3.3
gTTS==2.3.1
docker
duckduckgo-search
google-api-python-client #(https://developers.google.com/custom-search/v1/overview)
pinecone-client==2.2.1
redis
orjson
Pillow
coverage
numpy
gitpython==3.1.31
tweepy
jsonschema
flask
google-cloud-firestore
google-cloud-storage
google-cloud-datastore
gunicorn
Flask-Limiter==3.3.0
firebase-admin
